feat: empty settings form for streaming payment
